Fix install arch script

+++ .chezmoitemplates/install_arch.sh
@@ -2,6 +2,59 @@ RED='\033[0;31m'
 BLUE='\033[0;34m'
 NC='\033[0m' # No Color
 
+sudo pacman -Syu --noconfirm
+
+if ! command -v paru &>/dev/null; then
+  TEMP_FOLDER=$(mktemp -d)
+  sudo pacman -S --needed base-devel
+  git -C "$TEMP_FOLDER" clone --depth 1 https://aur.archlinux.org/paru.git
+  (
+    cd "$TEMP_FOLDER/paru"
+    makepkg -si
+  )
+  rm -rf "$TEMP_FOLDER"
+fi
+
+if grep -q "#" /etc/pacman.conf; then
+  sudo bash -c "cat >/etc/pacman.conf" <<EOF
+[options]
+CacheDir = /var/cache/pacman/pkg/
+HoldPkg = pacman glibc
+Architecture = auto
+Color
+ParallelDownloads = 10
+CheckSpace
+ILoveCandy
+VerbosePkgLists
+SigLevel = Required DatabaseOptional
+LocalFileSigLevel = Optional
+
+[core]
+Include = /etc/pacman.d/mirrorlist
+
+[extra]
+Include = /etc/pacman.d/mirrorlist
+
+[community]
+Include = /etc/pacman.d/mirrorlist
+
+[multilib]
+Include = /etc/pacman.d/mirrorlist
+EOF
+fi
+
+if ! grep -q "\[chaotic-aur\]" /etc/pacman.conf; then
+  sudo pacman-key --recv-key FBA220DFC880C036 --keyserver keyserver.ubuntu.com
+  sudo pacman-key --lsign-key FBA220DFC880C036
+  sudo pacman -U --noconfirm --needed 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-keyring.pkg.tar.zst' 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-mirrorlist.pkg.tar.zst'
+
+  sudo tee -a /etc/pacman.conf >/dev/null <<EOF
+
+[chaotic-aur]
+Include = /etc/pacman.d/chaotic-mirrorlist
+EOF
+fi
+
 # desktop enviroment
 dePackageList="
 ark \
@@ -64,7 +117,7 @@ packageList=(
   clang
   cmake
   code
-  conky-all
+  conky
   crudini
   curl
   exa
@@ -73,35 +126,32 @@ packageList=(
   git
   gparted
   imagemagick
-  jmtpfs
   jq
   keepassxc
   make
   micro
   mpv
+  mtpfs
   nano
-  ninja-build
+  ninja
   numlockx
   p7zip
-  python-is-python3
-  python3
-  python3-venv
+  python
   sshfs
-  systemd-zram-generator
-  texlive-fonts-recommended
-  texlive-lang-portuguese
-  texlive-latex-base
-  texlive-latex-extra
-  texlive-latex-recommended
+  starship
+  texlive-bibtexextra
+  texlive-bin
+  texlive-latexextra
+  texlive-latexindent-meta
   texlive-pictures
   texlive-pstricks
   texlive-science
-  texlive-xetex
   tmux
   ufw
   unzip
   xclip
   zip
+  zram-generator
 )
 
 if ! { [ -d /usr/share/plasma/plasmoids/org.kde.windowbuttons ] && [ -d /usr/share/plasma/plasmoids/org.kde.windowappmenu ]; }; then
@@ -111,37 +161,33 @@ if ! { [ -d /usr/share/plasma/plasmoids/org.kde.windowbuttons ] && [ -d /usr/sha
   )
 fi
 
-if [ -d "/sys/class/power_supply" ]; then
-  packageList=(
-    "${packageList[@]}"
-    tlp
-    tlp-rdw
-  )
+if [ -d "/proc/acpi/button/lid" ]; then
+  if ! pacman -Qi power-profiles-daemon &>/dev/null; then
+    packageList=(
+      "${packageList[@]}"
+      tlp
+      tlp-rdw
+    )
+  fi
 fi
 
 for package in "${packageList[@]}"; do
-  if ! dpkg -l "$package" &>/dev/null; then
-    sudo apt install -y "$package" || echo "${RED}Package ${BLUE}$package ${RED}not found!${NC}"
+  if ! pacman -Qi $package &>/dev/null; then
+    paru -S --noconfirm --needed --skipreview --nouseask --sudoloop $package ||
+      echo -e "${RED}Package(s) \"${BLUE}$package${RED}\" not found!${NC}"
   fi
 done
-
-if ! command -v bat &>/dev/null; then
-  mkdir -p ~/.local/bin
-  ln -s /usr/bin/batcat $HOME/.local/bin/bat
-fi
-
-if ! command -v starship &>/dev/null; then
-  curl -sS https://starship.rs/install.sh | sh -s -- -y
-fi
 
 if ! systemctl list-unit-files --state=enabled | grep ufw &>/dev/null; then
   sudo systemctl enable --now ufw
 fi
 
-if [ -d "/sys/class/power_supply" ]; then
-  if ! systemctl list-unit-files --state=enabled | grep tlp &>/dev/null; then
+if [ -d "/proc/acpi/button/lid" ]; then
+  if pacman -Qi tlp &>/dev/null && ! systemctl list-unit-files --state=enabled | grep tlp &>/dev/null; then
     sudo systemctl enable --now tlp
     sudo systemctl enable --now NetworkManager-dispatcher.service
     sudo systemctl mask systemd-rfkill.service systemd-rfkill.socket
   fi
 fi
+
+pacman -Qtdq | sudo pacman -Rns --noconfirm 2>/dev/null
